Working on profile with firebase connection
